fix: Made crystallize button appear with correct skill
The crystallize button in the inventory was appearing when a character had the Create Item skill, rather than the Crystallize skill.
This was due to a flag being set in the UserInfo packet based on L2PcInstance#hasDwarvenCraft().
Introduced a canCrystallize() method on L2PcInstance to drive the flag instead.