fix(raidboss): Fixed RB Eilhalder von Hellmann not spawning
When the spawns were loaded during ig daytime, the night RB would not be added to the list of bosses managed by the DayNightSpawnManager.
This bug was introduced when the backing map for night bosses was moved to a ConcurrentHashMap, which does not support null values.
Changed the map implementation to a HashMap as concurrent access is highly unlikely, which allows to initialise the boss status with null.