Added loot distribution features to party options.
In a party, adena is distributed evenly to all party members in the
range of the looter. This patch introduces three new config options
which can enable the same for all stackable items and/or all non
stackable items and/or specific item ids. The new options can be found
inside character.properties. You can read their description in the
properties file to have a more detailed description:
PartyEvenlyDistributeAllStackableItems
PartyEvenlyDistributeAllOtherItems
PartyEvenlyDistributeItems