Unhardcoding Officer Tolonis AI
Implemented Learn Skill Requested event.
Using Java records to represent events.
Removed `ON_` and `On` from event type and event object.
The `onLearnSkillRequested` method will receive the event object instead
of several parameters.
Methods to reference NPCs to the event will start with `bind` instead of
`add`.
A `setLearnSkillRequested(...)` method was avoided.

The rest of the changes to existing events will be done in a separated
commit.